Add solutions for common problems
As a new user I followed the steps in the README.md and fell into two problems before I could run Terraform plan.  Adding this to the documentation should help users to set the correct parameters.

Question: Shouldn't DevEx of the tool not guide users to the correct solution?
Maybe making the provider name a required parameter and setting strip to default? 
I don't know if that would add some negative effects elsewhere so that's why I'm adding the docs so that people like me would make fewer mistakes ;)